+++ Molten.Math/Vectors/Vector4UI.cs
@@ -241,10 +241,10 @@ namespace Molten
                 float inverse = 1.0F / length;
                 return new Vector4UI()
                 {
-			        X = (this.X * inverse),
-			        Y = (this.Y * inverse),
-			        Z = (this.Z * inverse),
-			        W = (this.W * inverse),
+			        X = (uint)(this.X * inverse),
+			        Y = (uint)(this.Y * inverse),
+			        Z = (uint)(this.Z * inverse),
+			        W = (uint)(this.W * inverse),
                 };
             }
             else
@@ -1060,10 +1060,10 @@ namespace Molten
 
             return new Vector4UI()
             {
-				X = (vector.X - ((2.0F * dot) * normal.X)),
-				Y = (vector.Y - ((2.0F * dot) * normal.Y)),
-				Z = (vector.Z - ((2.0F * dot) * normal.Z)),
-				W = (vector.W - ((2.0F * dot) * normal.W)),
+				X = (uint)(vector.X - ((2.0F * dot) * normal.X)),
+				Y = (uint)(vector.Y - ((2.0F * dot) * normal.Y)),
+				Z = (uint)(vector.Z - ((2.0F * dot) * normal.Z)),
+				W = (uint)(vector.W - ((2.0F * dot) * normal.W)),
             };
         }
